Rework history/task
In practice, almost all the tasks I'm creating have the same format of a
symbol name and a function with args, and the duplication feels a little
silly. I'm keeping the API for history.task/task, but in the history
namespace I'm calling it task-call and catch-task-call (for symmetry
with future-call). New `task` and `catch-task` macros.